requirements.txt
requests[socks]==2.32.3
psutil==6.0.0
ttkbootstrap==1.10.1
pillow==10.4.0